fix : Relocated components in accurate folders, fixed wrong class attribute to className attribute, fixed function name
