+++ src/pages/Main/MainSection/MainSection.scss
@@ -1,4 +1,4 @@
-@import '../../styles/constant';
+@import '../../../styles/constant';
 
 .MainSection {
   border: 1px $border solid;
