+++ src/pages/Main/Maincard/Maincard.scss
@@ -1,4 +1,4 @@
-@import '../../styles/constant.scss';
+@import '../../../styles/constant.scss';
 
 .MainCard {
   color: $text;
