+++ src/pages/ProductList/ProductList.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const ProductList = () => {
+  return (
+    <div className="ProductList">
+      <header>
+        <div className="breadCrumbs">Home / Spring</div>
+        <h1>Spring</h1>
+        <button type="button">Type: </button>
+      </header>
+    </div>
+  );
+};
+
+export default ProductList;
