+++ src/styles/constant.scss
@@ -1,6 +1,6 @@
 $bg-dark: #595959;
 $text: #333;
-$bg-light: #f5f8f6;
+$bg-light: #f7f8f6;
 $border: #ebebeb;
 
 @mixin center {
